python/rag/requirements.txt
galileo>=0.2.3
rich>=13.7.0
questionary>=2.0.1
openai==1.61.1
pydantic==2.10.6
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
certifi==2025.1.31
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
propcache==0.2.1
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
faiss-cpu>=1.7.4
sentence-transformers>=2.5.1
numpy>=1.24.0  # Required by FAISS and sentence-transformers
